Hold a total of postings in a type that has no ceiling
Two legal 201s of 9e18 put an account past int64 and every read endpoint
answered 500 from then on. A reversing entry recovered /v1/accounts and
could never recover /v1/trial-balance: `debits` is a filtered sum that
only ever grows, the book is append-only, and no entry a person could
write brought it back.

The book was never wrong. `amount_minor` is bigint and bounds one
posting; Postgres answers sum(bigint) with numeric and held the right
number throughout. Only the scan type could not read it.

- Added ledger.Minor, an arbitrary-precision integer that scans a numeric
  and marshals as a JSON number, and gave it Balance.AmountMinor and the
  three totals on Trial
- Widened the wire bodies over those, so balance_minor, debits_minor,
  credits_minor and net_minor carry what the database holds
- Widened `net` to numeric in assert_transaction_balances, so an entry
  whose legs sum past bigint is refused for not balancing and told by how
  much, rather than raising 22003 for a bare 422 that broke decision 14
- Took the net of an entry as a Minor in Entry.net and
  transactionRequest.net, so the number a refusal hands back is the one
  the entry carried
- Kept Leg.AmountMinor an int64: one posting is a bigint and a total of
  them is not, which is now what the Go types say
- Added decision 15 to DESIGN.md with the two routes rejected, the
  measurement that rejects a write-time ceiling, and the one statement a
  database already holding the schema needs
- Added three negative controls capping each total back at bigint

An existing book past the ceiling needs no migration: it starts answering
again as soon as the new binary runs.

+++ internal/schema/0001_ledger.sql
@@ -56,7 +56,8 @@ CREATE FUNCTION assert_transaction_balances(txn_id uuid) RETURNS void
 LANGUAGE plpgsql AS $$
 DECLARE
     legs bigint;
-    net  bigint;
+    -- numeric, not bigint: two legs the column will each hold can sum past one.
+    net  numeric;
 BEGIN
     SELECT count(*), coalesce(sum(amount_minor), 0)
       INTO legs, net
